Moved old script to  package name that indicates its deprecated
