+++ maptool/src/main/resources/net/rptools/maptool/client/script_deprecated/js/rptools.test.js
